[refactor] dto 폴더 위치 변경 (#17)
