+++ hr/src/main/java/com/reverse/hr/internal/dto/response/LoginUserProfileDTO.java
@@ -1,4 +1,4 @@
-package com.reverse.hr.internal.application.dto.response;
+package com.reverse.hr.internal.dto.response;
 
 /**
  * 로그인 성공 시 클라이언트에 내려주는 사용자 프로필 요약 정보 DTO.
