+++ hr/src/main/java/com/reverse/hr/internal/web/AuthController.java
@@ -3,10 +3,10 @@ package com.reverse.hr.internal.web;
 import com.reverse.core.exception.UnauthorizedException;
 import com.reverse.core.response.ApiResponse;
 import com.reverse.hr.internal.application.AuthService;
-import com.reverse.hr.internal.application.dto.request.ChangePasswordRequestDTO;
-import com.reverse.hr.internal.application.dto.request.InitializeRequestDTO;
-import com.reverse.hr.internal.application.dto.request.LoginRequestDTO;
-import com.reverse.hr.internal.application.dto.response.LoginResponseDTO;
+import com.reverse.hr.internal.dto.request.ChangePasswordRequestDTO;
+import com.reverse.hr.internal.dto.request.InitializeRequestDTO;
+import com.reverse.hr.internal.dto.request.LoginRequestDTO;
+import com.reverse.hr.internal.dto.response.LoginResponseDTO;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
